% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes_formatter.R
\name{create_child_parent_viewer}
\alias{create_child_parent_viewer}
\title{A Bn Model Viewer Function}
\usage{
create_child_parent_viewer(bn_model, delimiter = "|")
}
\arguments{
\item{bn_model}{& delimiter}
}
\value{
child-parent dataframe
}
\description{
This function create a child-parent dataframe of bn model nodes.
}
\keyword{delimiter}
