% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes_formatter.R
\name{create_from_to_viewer}
\alias{create_from_to_viewer}
\title{A Bn Model Viewer Function}
\usage{
create_from_to_viewer(bn_model)
}
\arguments{
\item{bn}{model}
}
\value{
from-to dataframe
create_from_to_viewer()
}
\description{
This function create a from-to dataframe of bn model nodes.
}
